requirements.txt
boto3
streamlit
streamlit-option-menu
re
s3fs
FilesConnection